+++ API/src/main/java/net/opmasterleo/packetuxui/PacketUxUiAPI.java
@@ -24,7 +24,7 @@ import net.opmasterleo.packetuxui.service.MenuService;
 
 public final class PacketUxUiAPI {
 
-    public static final String VERSION = "0.13.3";
+    public static final String VERSION = "0.13.4";
 
     private static final AtomicInteger RETAIN = new AtomicInteger();
     private static final Set<JavaPlugin> CLIENTS = new LinkedHashSet<>();